Moved ScanRequest to common data contract.
