+++ src/scanner/common/data-contract.ts
@@ -1,0 +1,7 @@
+export interface ScanRequest {
+    id: string;
+    name: string;
+    baseUrl: string;
+    scanUrl: string;
+    serviceTreeId: string;
+}
