+++ src/scanner/crawl-url/simple-crawler.ts
@@ -26,10 +26,3 @@ export interface CrawlRequest {
     baseUrl: string;
     serviceTreeId: string;
 }
-export interface ScanRequest {
-    id: string;
-    name: string;
-    baseUrl: string;
-    scanUrl: string;
-    serviceTreeId: string;
-}
